getSuperManager.sh, getTemporada.sh: - movida la variable de repositorio al fichero en sysconfig - control de existencia de la variable del repositorio - los scripts son mucho más parecidos ahora

+++ getSuperManager.sh
@@ -1,15 +1,21 @@
-#!/bin/bash
+#!/bin/bash 
 
 CONFIGFILE=/etc/sysconfig/SuperManager
 
-REPOURL="git@github.com:cesaralba/jimenezIntelligence.git"
 BASEDIR=$(cd "$(dirname $(readlink -e $0))/../" && pwd )
 WRKDIR="${BASEDIR}/wrk"
 TODAY=$(date '+%Y%m%d%H%M')
 
+[ -f ${CONFIGFILE} ] && source ${CONFIGFILE}
+
+if [ "x${SM_REPO}" = "x" ]
+then
+  echo "ORROR: No se ha suministrado valor para SM_REPO. Adios."
+  exit 1
+fi
 [ -d ${WRKDIR} ] && rm -rf  ${WRKDIR}
 mkdir -p ${WRKDIR}
-git clone -q --branch master ${REPOURL} ${WRKDIR}
+git clone -q --branch master ${SM_REPO} ${WRKDIR}
 
 if [ $? != 0 ]
 then
@@ -18,8 +24,6 @@ then
 fi
 
 source /home/calba/devel/SuperManagerPython/SACBenv/bin/activate
-
-[ -f ${CONFIGFILE} ] && source ${CONFIGFILE}
 
 ORIGSMFILE="${BASEDIR}/datos/SM2017.latest.p"
 DESTSMFILE="${BASEDIR}/datos/SM2017.newest.p"
@@ -36,7 +40,6 @@ then
     mv ${DESTSMFILE} ${ORIGSMFILE}
   fi
 fi  
-
 
 chown -R calba /home/calba/devel/SuperManager
